shopping-cart/requirements.txt
Flask==2.2.0
requests==2.25.1
mariadb==1.0.9
flask_sqlalchemy==3.0.2
protobuf==3.20.1
prometheus-flask-exporter==0.22.3
opentelemetry-api==1.11.1
opentelemetry-sdk==1.11.1
opentelemetry-exporter-otlp==1.11.1
opentelemetry-instrumentation-flask==0.30b1
opentelemetry-instrumentation-logging==0.30b1
opentelemetry-instrumentation-sqlalchemy==0.30b1
opentelemetry-instrumentation-wsgi==0.30b1
uwsgi==2.0.20